requirements.txt
streamlit
pandas
requests
sentence-transformers
torch
openai
python-dotenv